Move SynapseClientError to its own file
