+++ src/lib/containers/MarkdownSynapse.tsx
@@ -7,7 +7,7 @@ import SynapseImage from './widgets/SynapseImage'
 import SynapsePlot from './widgets/SynapsePlot'
 import SynapseVideo from './widgets/SynapseVideo'
 import { ErrorBanner } from './ErrorBanner'
-import { SynapseClientError } from '../utils/SynapseClient'
+import { SynapseClientError } from '../utils/SynapseClientError'
 import { Button } from 'react-bootstrap'
 import { SynapseContext } from '../utils/SynapseContext'
 import IDUReport from './IDUReport'
