+++ src/lib/containers/entity/annotations/SchemaDrivenAnnotationEditor.tsx
@@ -15,7 +15,7 @@ import {
 } from '../../../utils/hooks/SynapseAPI/entity/useEntity'
 import { useGetSchemaBinding } from '../../../utils/hooks/SynapseAPI/entity/useEntityBoundSchema'
 import { useGetSchema } from '../../../utils/hooks/SynapseAPI/useSchema'
-import { SynapseClientError } from '../../../utils/SynapseClient'
+import { SynapseClientError } from '../../../utils/SynapseClientError'
 import { EntityJson, entityJsonKeys } from '../../../utils/synapseTypes'
 import { SynapseSpinner } from '../../LoadingScreen'
 import { AdditionalPropertiesSchemaField } from './AdditionalPropertiesSchemaField'
