+++ src/lib/containers/home_page/goals/Goals.tsx
@@ -6,7 +6,8 @@ import {
   BatchFileRequest,
 } from '../../../utils/synapseTypes'
 import { SynapseConstants } from '../../../utils'
-import { SynapseClientError, getFiles } from '../../../utils/SynapseClient'
+import { getFiles } from '../../../utils/SynapseClient'
+import { SynapseClientError } from '../../../utils/SynapseClientError'
 import { ErrorBanner } from '../../ErrorBanner'
 import useGetQueryResultBundle from '../../../utils/hooks/SynapseAPI/entity/useGetQueryResultBundle'
 import useShowDesktop from '../../../utils/hooks/useShowDesktop'
