+++ src/lib/utils/hooks/SynapseAPI/entity/useEntity.ts
@@ -16,7 +16,7 @@ import {
   UseQueryOptions,
 } from 'react-query'
 import { SynapseClient } from '../../..'
-import { SynapseClientError } from '../../../SynapseClient'
+import { SynapseClientError } from '../../../SynapseClientError'
 import { useSynapseContext } from '../../../SynapseContext'
 import {
   Entity,
