+++ src/lib/utils/hooks/SynapseAPI/entity/useEntityBoundSchema.ts
@@ -1,6 +1,6 @@
 import { UseQueryOptions, useQuery } from 'react-query'
 import { SynapseClient } from '../../..'
-import { SynapseClientError } from '../../../SynapseClient'
+import { SynapseClientError } from '../../../SynapseClientError'
 import { useSynapseContext } from '../../../SynapseContext'
 import { JsonSchemaObjectBinding } from '../../../synapseTypes/Schema/JsonSchemaObjectBinding'
 import { ValidationResults } from '../../../synapseTypes/Schema/ValidationResults'
