+++ src/lib/utils/hooks/SynapseAPI/entity/useGetEntityChildren.ts
@@ -7,7 +7,7 @@ import {
   UseQueryOptions,
 } from 'react-query'
 import { SynapseClient } from '../../..'
-import { SynapseClientError } from '../../../SynapseClient'
+import { SynapseClientError } from '../../../SynapseClientError'
 import { useSynapseContext } from '../../../SynapseContext'
 import {
   EntityChildrenRequest,
