+++ src/lib/utils/hooks/SynapseAPI/useGetDownloadListActionsRequired.ts
@@ -5,7 +5,7 @@ import {
   UseQueryOptions,
 } from 'react-query'
 import { SynapseClient } from '../..'
-import { SynapseClientError } from '../../SynapseClient'
+import { SynapseClientError } from '../../SynapseClientError'
 import { useSynapseContext } from '../../SynapseContext'
 import { ActionRequiredRequest } from '../../synapseTypes/DownloadListV2/QueryRequestDetails'
 import { ActionRequiredResponse } from '../../synapseTypes/DownloadListV2/QueryResponseDetails'
