+++ src/lib/utils/hooks/SynapseAPI/useGetDownloadListStatistics.ts
@@ -1,6 +1,6 @@
 import { useQuery, UseQueryOptions } from 'react-query'
 import { SynapseClient } from '../..'
-import { SynapseClientError } from '../../SynapseClient'
+import { SynapseClientError } from '../../SynapseClientError'
 import { useSynapseContext } from '../../SynapseContext'
 import { FilesStatisticsResponse } from '../../synapseTypes/DownloadListV2/QueryResponseDetails'
