+++ src/lib/utils/hooks/SynapseAPI/useGetUserChallenges.ts
@@ -1,6 +1,6 @@
 import { useInfiniteQuery, UseInfiniteQueryOptions } from 'react-query'
 import { SynapseClient } from '../..'
-import { SynapseClientError } from '../../SynapseClient'
+import { SynapseClientError } from '../../SynapseClientError'
 import { useSynapseContext } from '../../SynapseContext'
 import { ChallengeWithProjectHeaderPagedResults } from '../../synapseTypes/ChallengePagedResults'
