+++ src/lib/utils/hooks/SynapseAPI/useProjects.ts
@@ -5,7 +5,7 @@ import {
   UseQueryOptions,
 } from 'react-query'
 import { SynapseClient } from '../..'
-import { SynapseClientError } from '../../SynapseClient'
+import { SynapseClientError } from '../../SynapseClientError'
 import { useSynapseContext } from '../../SynapseContext'
 import { ProjectHeaderList } from '../../synapseTypes'
 import { GetProjectsParameters } from '../../synapseTypes/GetProjectsParams'
