+++ src/lib/utils/hooks/SynapseAPI/useSearch.ts
@@ -6,7 +6,7 @@ import {
   UseQueryOptions,
 } from 'react-query'
 import { SynapseClient } from '../..'
-import { SynapseClientError } from '../../SynapseClient'
+import { SynapseClientError } from '../../SynapseClientError'
 import { useSynapseContext } from '../../SynapseContext'
 import { SearchQuery, SearchResults } from '../../synapseTypes/Search'
