+++ src/lib/utils/hooks/SynapseAPI/useUserBundle.ts
@@ -1,6 +1,6 @@
 import { useQuery, UseQueryOptions } from 'react-query'
 import { SynapseClient } from '../..'
-import { SynapseClientError } from '../../SynapseClient'
+import { SynapseClientError } from '../../SynapseClientError'
 import { useSynapseContext } from '../../SynapseContext'
 import { NotificationEmail, UserBundle, UserProfile } from '../../synapseTypes'
 import {
